requirements.txt
docker-py>=1.0.0
py-dateutil